rebrand, remove multisrc
