+++ src/all/galaxy/build.gradle
@@ -1,0 +1,7 @@
+ext {
+    extName = 'Galaxy'
+    extClass = '.GalaxyFactory'
+    extVersionCode = 1
+}
+
+apply from: "$rootDir/common.gradle"
