--- src/ar/galaxymanga/src/eu/kanade/tachiyomi/extension/ar/galaxymanga/GalaxyManga.kt
@@ -1,11 +1,0 @@
-package eu.kanade.tachiyomi.extension.ar.galaxymanga
-
-import eu.kanade.tachiyomi.multisrc.flixscans.FlixScans
-
-class GalaxyManga : FlixScans(
-    "جالاكسي مانجا",
-    "https://galaxymanga.net",
-    "ar",
-) {
-    override val versionId = 2
-}
