--- src/en/flixscans/src/eu/kanade/tachiyomi/extension/en/flixscans/FlixScans.kt
@@ -1,9 +1,0 @@
-package eu.kanade.tachiyomi.extension.en.flixscans
-
-import eu.kanade.tachiyomi.multisrc.flixscans.FlixScans
-
-class FlixScans : FlixScans(
-    "Flix Scans",
-    "https://flixscans.net",
-    "en",
-)
